resparser/pyproject.toml
[project]
name = "resparser"
version = "0.1.0"
description = "Default template for PDM package"


authors = [
    {name = "Zhu, Mingang", email = "fzhurd@gmail.com"},
]

#[tool.pdm.source]
#pypi = { url = "https://pypi.org/simple/", verify_ssl = true }

dependencies = ["openai>=2.14.0", "google-generativeai>=0.8.6", "pytest>=9.0.2", "pdfplumber>=0.11.8", "python-docx>=1.2.0"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple/"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
